Added a better fight view

+++ Arena-Battle/Game/Classes/Assassin.swift
@@ -1,17 +1,17 @@
 class Assassin: Fighter {
-    var energy: Int = 0
-    var agility: Int = 0
+    var energy: Float = 0
+    var agility: Float = 0
     
-    init(name: String, health: Int, maxAttackPower: Int, maxBlock: Int, energy: Int, agility: Int) {
+    init(name: String, health: Float, maxAttackPower: Float, maxBlock: Float, energy: Float, agility: Float) {
         super.init(name, health, maxAttackPower, maxBlock)
         self.agility = agility
         self.energy = energy
     }
     
-    override func block() -> Int {
-        let randomDodgeAmount = Int.random(in: 1...100)
+    override func block() -> Float {
+        let randomDodgeAmount = Float.random(in: 1...100)
         if randomDodgeAmount < self.agility {
-            return 10000
+            return 10000.0
         } else {
             return super.block()
         }
