+++ Arena-Battle/Game/Classes/Mage.swift
@@ -1,7 +1,7 @@
 class Mage: Fighter {
-    var mana: Int = 0
+    var mana: Float = 0
     
-    init(name: String, health: Int, maxAttackPower: Int, maxBlock: Int, mana: Int) {
+    init(name: String, health: Float, maxAttackPower: Float, maxBlock: Float, mana: Float) {
         super.init(name, health, maxAttackPower, maxBlock)
         self.mana = mana
     }
